+++ bitcoin/src/psbt/input.rs
@@ -1,23 +1,16 @@
 // SPDX-License-Identifier: CC0-1.0
 
+//! A PSBT input.
+
 use hashes::{hash160, ripemd160, sha256, sha256d};
 use secp256k1::XOnlyPublicKey;
 
-use super::Map;
+use super::{map, raw};
 use crate::bip32::KeySource;
 use crate::crypto::key::PublicKey;
 use crate::crypto::{ecdsa, taproot};
-use crate::prelude::{btree_map, BTreeMap, Borrow, Box, Vec};
-use crate::psbt::consts::{
-    PSBT_IN_BIP32_DERIVATION, PSBT_IN_FINAL_SCRIPTSIG, PSBT_IN_FINAL_SCRIPTWITNESS,
-    PSBT_IN_HASH160, PSBT_IN_HASH256, PSBT_IN_NON_WITNESS_UTXO, PSBT_IN_PARTIAL_SIG,
-    PSBT_IN_PROPRIETARY, PSBT_IN_REDEEM_SCRIPT, PSBT_IN_RIPEMD160, PSBT_IN_SHA256,
-    PSBT_IN_SIGHASH_TYPE, PSBT_IN_TAP_BIP32_DERIVATION, PSBT_IN_TAP_INTERNAL_KEY,
-    PSBT_IN_TAP_KEY_SIG, PSBT_IN_TAP_LEAF_SCRIPT, PSBT_IN_TAP_MERKLE_ROOT, PSBT_IN_TAP_SCRIPT_SIG,
-    PSBT_IN_WITNESS_SCRIPT, PSBT_IN_WITNESS_UTXO,
-};
-use crate::psbt::error::PsbtHash;
-use crate::psbt::{raw, PsbtSighashType, Error};
+use crate::prelude::{BTreeMap, Vec};
+use crate::psbt::PsbtSighashType;
 use crate::script::ScriptBuf;
 use crate::sighash::{
     EcdsaSighashType, InvalidSighashTypeError, NonStandardSighashTypeError,
@@ -36,66 +29,144 @@ pub struct Input {
     /// `Option::Some` for inputs which spend non-segwit outputs or
     /// if it is unknown whether an input spends a segwit output.
     pub non_witness_utxo: Option<Transaction>,
+
     /// The transaction output this input spends from. Should only be
     /// `Option::Some` for inputs which spend segwit outputs,
     /// including P2SH embedded ones.
     pub witness_utxo: Option<TxOut>,
+
     /// A map from public keys to their corresponding signature as would be
     /// pushed to the stack from a scriptSig or witness for a non-Taproot inputs.
     pub partial_sigs: BTreeMap<PublicKey, ecdsa::Signature>,
+
     /// The sighash type to be used for this input. Signatures for this input
     /// must use the sighash type.
     pub sighash_type: Option<PsbtSighashType>,
+
     /// The redeem script for this input.
     pub redeem_script: Option<ScriptBuf>,
+
     /// The witness script for this input.
     pub witness_script: Option<ScriptBuf>,
+
     /// A map from public keys needed to sign this input to their corresponding
     /// master key fingerprints and derivation paths.
     #[cfg_attr(feature = "serde", serde(with = "crate::serde_utils::btreemap_as_seq"))]
     pub bip32_derivation: BTreeMap<secp256k1::PublicKey, KeySource>,
+
     /// The finalized, fully-constructed scriptSig with signatures and any other
     /// scripts necessary for this input to pass validation.
     pub final_script_sig: Option<ScriptBuf>,
+
     /// The finalized, fully-constructed scriptWitness with signatures and any
     /// other scripts necessary for this input to pass validation.
     pub final_script_witness: Option<Witness>,
+
     /// RIPEMD160 hash to preimage map.
     #[cfg_attr(feature = "serde", serde(with = "crate::serde_utils::btreemap_byte_values"))]
     pub ripemd160_preimages: BTreeMap<ripemd160::Hash, Vec<u8>>,
+
     /// SHA256 hash to preimage map.
     #[cfg_attr(feature = "serde", serde(with = "crate::serde_utils::btreemap_byte_values"))]
     pub sha256_preimages: BTreeMap<sha256::Hash, Vec<u8>>,
+
     /// HSAH160 hash to preimage map.
     #[cfg_attr(feature = "serde", serde(with = "crate::serde_utils::btreemap_byte_values"))]
     pub hash160_preimages: BTreeMap<hash160::Hash, Vec<u8>>,
+
     /// HAS256 hash to preimage map.
     #[cfg_attr(feature = "serde", serde(with = "crate::serde_utils::btreemap_byte_values"))]
     pub hash256_preimages: BTreeMap<sha256d::Hash, Vec<u8>>,
+
     /// Serialized Taproot signature with sighash type for key spend.
     pub tap_key_sig: Option<taproot::Signature>,
+
     /// Map of `<xonlypubkey>|<leafhash>` with signature.
     #[cfg_attr(feature = "serde", serde(with = "crate::serde_utils::btreemap_as_seq"))]
     pub tap_script_sigs: BTreeMap<(XOnlyPublicKey, TapLeafHash), taproot::Signature>,
+
     /// Map of Control blocks to Script version pair.
     #[cfg_attr(feature = "serde", serde(with = "crate::serde_utils::btreemap_as_seq"))]
     pub tap_scripts: BTreeMap<ControlBlock, (ScriptBuf, LeafVersion)>,
+
     /// Map of tap root x only keys to origin info and leaf hashes contained in it.
     #[cfg_attr(feature = "serde", serde(with = "crate::serde_utils::btreemap_as_seq"))]
     pub tap_key_origins: BTreeMap<XOnlyPublicKey, (Vec<TapLeafHash>, KeySource)>,
+
     /// Taproot Internal key.
     pub tap_internal_key: Option<XOnlyPublicKey>,
+
     /// Taproot Merkle root.
     pub tap_merkle_root: Option<TapNodeHash>,
+
     /// Proprietary key-value pairs for this input.
     #[cfg_attr(feature = "serde", serde(with = "crate::serde_utils::btreemap_as_seq_byte_values"))]
     pub proprietary: BTreeMap<raw::ProprietaryKey, Vec<u8>>,
+
     /// Unknown key-value pairs for this input.
     #[cfg_attr(feature = "serde", serde(with = "crate::serde_utils::btreemap_as_seq_byte_values"))]
     pub unknown: BTreeMap<raw::Key, Vec<u8>>,
 }
 
 impl Input {
+    /// Creates an `Input` from a serializable [`Input`].
+    ///
+    /// [`Input`]: crate::psbt::map::Input
+    pub fn from_serializable(input: map::Input) -> Self {
+        Input {
+            non_witness_utxo: input.non_witness_utxo,
+            witness_utxo: input.witness_utxo,
+            partial_sigs: input.partial_sigs,
+            sighash_type: input.sighash_type,
+            redeem_script: input.redeem_script,
+            witness_script: input.witness_script,
+            bip32_derivation: input.bip32_derivation,
+            final_script_sig: input.final_script_sig,
+            final_script_witness: input.final_script_witness,
+            ripemd160_preimages: input.ripemd160_preimages,
+            sha256_preimages: input.sha256_preimages,
+            hash160_preimages: input.hash160_preimages,
+            hash256_preimages: input.hash256_preimages,
+            tap_key_sig: input.tap_key_sig,
+            tap_script_sigs: input.tap_script_sigs,
+            tap_scripts: input.tap_scripts,
+            tap_key_origins: input.tap_key_origins,
+            tap_internal_key: input.tap_internal_key,
+            tap_merkle_root: input.tap_merkle_root,
+            proprietary: input.proprietary,
+            unknown: input.unknown,
+        }
+    }
+
+    /// Converts this `Input` into a serializable [`Input`].
+    ///
+    /// [`Input`]: crate::psbt::map::Input
+    pub fn into_serializable(self) -> map::Input {
+        map::Input {
+            non_witness_utxo: self.non_witness_utxo,
+            witness_utxo: self.witness_utxo,
+            partial_sigs: self.partial_sigs,
+            sighash_type: self.sighash_type,
+            redeem_script: self.redeem_script,
+            witness_script: self.witness_script,
+            bip32_derivation: self.bip32_derivation,
+            final_script_sig: self.final_script_sig,
+            final_script_witness: self.final_script_witness,
+            ripemd160_preimages: self.ripemd160_preimages,
+            sha256_preimages: self.sha256_preimages,
+            hash160_preimages: self.hash160_preimages,
+            hash256_preimages: self.hash256_preimages,
+            tap_key_sig: self.tap_key_sig,
+            tap_script_sigs: self.tap_script_sigs,
+            tap_scripts: self.tap_scripts,
+            tap_key_origins: self.tap_key_origins,
+            tap_internal_key: self.tap_internal_key,
+            tap_merkle_root: self.tap_merkle_root,
+            proprietary: self.proprietary,
+            unknown: self.unknown,
+        }
+    } 
+
     /// Obtains the [`EcdsaSighashType`] for this input if one is specified. If no sighash type is
     /// specified, returns [`EcdsaSighashType::All`].
     ///
@@ -118,125 +189,6 @@ impl Input {
         self.sighash_type
             .map(|sighash_type| sighash_type.taproot_hash_ty())
             .unwrap_or(Ok(TapSighashType::Default))
-    }
-
-    pub(super) fn insert_pair(&mut self, pair: raw::Pair) -> Result<(), Error> {
-        let raw::Pair { key: raw_key, value: raw_value } = pair;
-
-        match raw_key.type_value {
-            PSBT_IN_NON_WITNESS_UTXO => {
-                impl_psbt_insert_pair! {
-                    self.non_witness_utxo <= <raw_key: _>|<raw_value: Transaction>
-                }
-            }
-            PSBT_IN_WITNESS_UTXO => {
-                impl_psbt_insert_pair! {
-                    self.witness_utxo <= <raw_key: _>|<raw_value: TxOut>
-                }
-            }
-            PSBT_IN_PARTIAL_SIG => {
-                impl_psbt_insert_pair! {
-                    self.partial_sigs <= <raw_key: PublicKey>|<raw_value: ecdsa::Signature>
-                }
-            }
-            PSBT_IN_SIGHASH_TYPE => {
-                impl_psbt_insert_pair! {
-                    self.sighash_type <= <raw_key: _>|<raw_value: PsbtSighashType>
-                }
-            }
-            PSBT_IN_REDEEM_SCRIPT => {
-                impl_psbt_insert_pair! {
-                    self.redeem_script <= <raw_key: _>|<raw_value: ScriptBuf>
-                }
-            }
-            PSBT_IN_WITNESS_SCRIPT => {
-                impl_psbt_insert_pair! {
-                    self.witness_script <= <raw_key: _>|<raw_value: ScriptBuf>
-                }
-            }
-            PSBT_IN_BIP32_DERIVATION => {
-                impl_psbt_insert_pair! {
-                    self.bip32_derivation <= <raw_key: secp256k1::PublicKey>|<raw_value: KeySource>
-                }
-            }
-            PSBT_IN_FINAL_SCRIPTSIG => {
-                impl_psbt_insert_pair! {
-                    self.final_script_sig <= <raw_key: _>|<raw_value: ScriptBuf>
-                }
-            }
-            PSBT_IN_FINAL_SCRIPTWITNESS => {
-                impl_psbt_insert_pair! {
-                    self.final_script_witness <= <raw_key: _>|<raw_value: Witness>
-                }
-            }
-            PSBT_IN_RIPEMD160 => {
-                psbt_insert_hash_pair! {
-                    &mut self.ripemd160_preimages <= raw_key|raw_value|ripemd160::Hash|PsbtHash::Ripemd
-                }
-            }
-            PSBT_IN_SHA256 => {
-                psbt_insert_hash_pair! {
-                    &mut self.sha256_preimages <= raw_key|raw_value|sha256::Hash|PsbtHash::Sha256
-                }
-            }
-            PSBT_IN_HASH160 => {
-                psbt_insert_hash_pair! {
-                    &mut self.hash160_preimages <= raw_key|raw_value|hash160::Hash|PsbtHash::Hash160
-                }
-            }
-            PSBT_IN_HASH256 => {
-                psbt_insert_hash_pair! {
-                    &mut self.hash256_preimages <= raw_key|raw_value|sha256d::Hash|PsbtHash::Hash256
-                }
-            }
-            PSBT_IN_TAP_KEY_SIG => {
-                impl_psbt_insert_pair! {
-                    self.tap_key_sig <= <raw_key: _>|<raw_value: taproot::Signature>
-                }
-            }
-            PSBT_IN_TAP_SCRIPT_SIG => {
-                impl_psbt_insert_pair! {
-                    self.tap_script_sigs <= <raw_key: (XOnlyPublicKey, TapLeafHash)>|<raw_value: taproot::Signature>
-                }
-            }
-            PSBT_IN_TAP_LEAF_SCRIPT => {
-                impl_psbt_insert_pair! {
-                    self.tap_scripts <= <raw_key: ControlBlock>|< raw_value: (ScriptBuf, LeafVersion)>
-                }
-            }
-            PSBT_IN_TAP_BIP32_DERIVATION => {
-                impl_psbt_insert_pair! {
-                    self.tap_key_origins <= <raw_key: XOnlyPublicKey>|< raw_value: (Vec<TapLeafHash>, KeySource)>
-                }
-            }
-            PSBT_IN_TAP_INTERNAL_KEY => {
-                impl_psbt_insert_pair! {
-                    self.tap_internal_key <= <raw_key: _>|< raw_value: XOnlyPublicKey>
-                }
-            }
-            PSBT_IN_TAP_MERKLE_ROOT => {
-                impl_psbt_insert_pair! {
-                    self.tap_merkle_root <= <raw_key: _>|< raw_value: TapNodeHash>
-                }
-            }
-            PSBT_IN_PROPRIETARY => {
-                let key = raw::ProprietaryKey::try_from(raw_key.clone())?;
-                match self.proprietary.entry(key) {
-                    btree_map::Entry::Vacant(empty_key) => {
-                        empty_key.insert(raw_value);
-                    }
-                    btree_map::Entry::Occupied(_) => return Err(Error::DuplicateKey(raw_key)),
-                }
-            }
-            _ => match self.unknown.entry(raw_key) {
-                btree_map::Entry::Vacant(empty_key) => {
-                    empty_key.insert(raw_value);
-                }
-                btree_map::Entry::Occupied(k) => return Err(Error::DuplicateKey(k.key().clone())),
-            },
-        }
-
-        Ok(())
     }
 
     /// Combines this [`Input`] with `other` `Input` (as described by BIP 174).
@@ -269,99 +221,6 @@ impl Input {
         combine!(tap_merkle_root, self, other);
     }
 }
-
-impl Map for Input {
-    fn get_pairs(&self) -> Vec<raw::Pair> {
-        let mut rv: Vec<raw::Pair> = Default::default();
-
-        impl_psbt_get_pair! {
-            rv.push(self.non_witness_utxo, PSBT_IN_NON_WITNESS_UTXO)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push(self.witness_utxo, PSBT_IN_WITNESS_UTXO)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push_map(self.partial_sigs, PSBT_IN_PARTIAL_SIG)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push(self.sighash_type, PSBT_IN_SIGHASH_TYPE)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push(self.redeem_script, PSBT_IN_REDEEM_SCRIPT)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push(self.witness_script, PSBT_IN_WITNESS_SCRIPT)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push_map(self.bip32_derivation, PSBT_IN_BIP32_DERIVATION)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push(self.final_script_sig, PSBT_IN_FINAL_SCRIPTSIG)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push(self.final_script_witness, PSBT_IN_FINAL_SCRIPTWITNESS)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push_map(self.ripemd160_preimages, PSBT_IN_RIPEMD160)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push_map(self.sha256_preimages, PSBT_IN_SHA256)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push_map(self.hash160_preimages, PSBT_IN_HASH160)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push_map(self.hash256_preimages, PSBT_IN_HASH256)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push(self.tap_key_sig, PSBT_IN_TAP_KEY_SIG)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push_map(self.tap_script_sigs, PSBT_IN_TAP_SCRIPT_SIG)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push_map(self.tap_scripts, PSBT_IN_TAP_LEAF_SCRIPT)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push_map(self.tap_key_origins, PSBT_IN_TAP_BIP32_DERIVATION)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push(self.tap_internal_key, PSBT_IN_TAP_INTERNAL_KEY)
-        }
-
-        impl_psbt_get_pair! {
-            rv.push(self.tap_merkle_root, PSBT_IN_TAP_MERKLE_ROOT)
-        }
-        for (key, value) in self.proprietary.iter() {
-            rv.push(raw::Pair { key: key.to_key(), value: value.clone() });
-        }
-
-        for (key, value) in self.unknown.iter() {
-            rv.push(raw::Pair { key: key.clone(), value: value.clone() });
-        }
-
-        rv
-    }
-}
-
-impl_psbtmap_ser_de_serialize!(Input);
 
 #[cfg(test)]
 mod test {
